fixed error

+++ twitter-to-kafka-common-config-data/src/main/java/com/app/t2k/kafka/to/elastic/config/KafkaConsumerConfigData.java
@@ -27,6 +27,6 @@ public class KafkaConsumerConfigData {
     private Integer maxPartitionFetchBytesDefault;
     private Integer maxPartitionFetchBytesBoostFactor;
     private Long pollTimeoutMs;
-    private Long maxPollRecords;
+    private Integer maxPollRecords;
 
 }
